In Slater.py: Refactored Slater.s() to handle any number of shells (up to 9). In other words, the input for this function can now be a list of any length between 1 and 9, instead of only exactly 9. (will eventually want to implement this better and have it be able to handle any number of shells we could conceivably be using, but it's fine for the moment)
In SlaterGUI.py, removed piece of code that copies only the first 9 elements of the electron config list,as this is no longer needed.

Renamed script Routines.py to GridStretch.py because we plan on removing everything but the functions that deal with exponential grid stretching, as everything other than that is not being used.

In script Atoms.py: Removed trailing zero's from jellium elements.