In Slater.py: Added a (guess) value to nStar dictionary, made various clean code edits, ie refactored function name, removed redundant variable declaration and unused print statements, removed redundant decimal point. In newDensity(), pulled more data from the atom object instead of an independent list, and removed the check that omits returned data for unpopulated electron orbitals, as this breaks my automatic legend population system in matplotlib.
In SlaterGUI.py, changed the way the legend is populated so that it pulls information directly from the atom object instead of from an external list. Note this is an independent and unrelated change from the above-mentioned edit to fix incorrect legend population originating in Slater.newDensity()

In Atoms.py,
Fixed a typo in a comment.