Added a variable to the newAtom object to store shielding values in the atom object itself.
Removed now deprecated function 'Slater.ComputeShieldingConstants()' which was the original method of computing shielding constants, but was not capable of handling atoms of arbitrary size. It has been replaced with 'newComputeShieldingConstants()' which will likely have its name changed to drop the 'new' in the future.

Also removed similarly deprecated 'density()' function in Slater.py, replaced with 'newDensity()'

Commented out all Slater.py functions that would throw errors if they were used with the newAtom object, will adapt any that are potentially useful to do so in the future.

Added new function 'computeTotalEnergy() that does exactly what it says on the tin. Should probably make it an object function/method since it only takes the atom object as input.'