In Atoms.py: Fixed an incorrect statement in a comment. Fixed incorrect calculation of total energy.
In inputFunctions.py:
extended maximum allowed electron occupation string from 10 orbitals to 100 (temporary, this is just an arbitrarily large number to effectively remove the limit until I refactor the whole function, as the length limit is (probably) no longer needed.)

Added an upper bound on the value in getAtomicNumber (used to get the atomic number of the atom to model in manual mode) due to concerns about internal zeros with the shell occupancy, and the resulting complexity of asking precisely which orbitals should be filled.

In Slater.py:
fixed incorrect algorithm in computeTotalEnergy()

In easygui.py:
Raised upperbound of integerbox() from 99 to 999. That way if someone enters something like 100 or 120 into inputFunctions.getAtomicNumber() it will show my notification which tells the user why their input was rejected, instead of just saying the input is out of bounds. I could forgo adding the logic in my code entirely and just set the upper bound to 88 instead, but that might have unintended consequences further down the line, although it would be unlikely. If I really wanted to, I think I could just manually define the upper bound for the specific instance of integerbox() I call, but again, that would give the user a less descriptive error message that would not say exactly why their input was rejected and also would not tell them to use the automatic mode for heavier elements.