In Atoms.py: Made the newAtom object compute its shieldingValues and totalEnergy in its initialization using internal functions, instead of the functions in Slater.py.
In Slater.py:
Removed newComputeShieldingConstants() and computeTotalEnergy() functions because these are now calculated with object functions during the initialization of the newAtom object.

In SlaterGUI.py:
Removed declaration of variable 'lists' which was calling Slater.newComputeShieldingConstants() as it was not being used anywhere.
Also moved the print statement that prints the total energy from the bottom of the script to shortly after the atom object is initialized, that way you don't have to close the plot in order to see it. (in case you needed to look at the plot and know what the total energy was at the same time)