In Slater.py: Removed unused print statement. Changed print statement that alerts the user of an unbound atom so that it gives you information on the implications of such a condition, rather than (incorrectly) saying "EXITING SOON".
In inputFunctions.py:
Added window titles to several input boxes as well as various message boxes that are used to deal with "soft errors" that the input functions can handle during runtime (namely, invalid inputs and/or parsing anomalies arising from invalid inputs)
Removed all checks/limitations for electron shell configuration input in getElectronConfigInput() related to a maximum length, as there is (theoretically) no upper limit on what the shielding constant calculation function can handle.
Added a new check and associated message box error that handles situations where orbitalConfigList[] has an element that is an empty string, even when the user input itself was not an empty string. (This is caused by having an input containing two or more consecutive periods without another character in between) Without this, the input was cleared but no message was displayed explaining why it was cleared.
Finally, Re-wrote the prompt for manual entry of the electron configuration to clarify that ALL orbitals inside the highest populated orbital must be included in the input, even if not every one of those orbitals itself is populated. (see below for more details)

In Atoms.py:
More thoroughly investigated the proper formatting of the electron orbitals with context to internal zeros. Changed nothing here, but added a commented link to a website that proved helpful.